--- a/src/ReinforcementLearningCore/src/policies/agent.jl
+++ b/src/ReinforcementLearningCore/src/policies/agent.jl
@@ -4,6 +4,23 @@ using Base.Threads: @spawn
 
 using Functors: @functor
 
+mutable struct AgentCache{A,S,R}
+    action::Union{Missing, A}
+    state::Union{Missing, S}
+    reward::Union{Missing, R}
+    terminal::Union{Missing, Bool}
+
+    function AgentCache(policy::AbstractPolicy, env::AbstractEnv)
+        new{typeof(policy(env)), typeof(state(env)), typeof(reward(env))}(missing, missing, missing)
+    end
+
+    function AgentCache()
+        new{Any, Any, Any}(missing, missing, missing, missing)
+    end
+end
+
+const cache_attrs = (:action, :state, :reward, :terminal)
+
 """
     Agent(;policy, trajectory) <: AbstractPolicy
 
@@ -16,10 +33,15 @@ passed to the policy.
 mutable struct Agent{P,T} <: AbstractPolicy
     policy::P
     trajectory::T
-    cache::NamedTuple # trajectory does not support partial inserting
+    cache::AgentCache # trajectory does not support partial inserting
 
-    function Agent(policy::P, trajectory::T, cache=NamedTuple()) where {P,T}
-        agent = new{P,T}(policy, trajectory, cache)
+    function Agent(policy::P, trajectory::T; env=missing) where {P,T}
+        if !ismissing(env)
+            agent = new{P,T}(policy, trajectory, AgentCache(policy, env))
+        else
+            agent = new{P,T}(policy, trajectory, AgentCache())
+        end
+
         if TrajectoryStyle(trajectory) === AsyncTrajectoryStyle()
             bind(trajectory, @spawn(optimise!(policy, trajectory)))
         end
@@ -27,7 +49,7 @@ mutable struct Agent{P,T} <: AbstractPolicy
     end
 end
 
-Agent(; policy, trajectory, cache=NamedTuple()) = Agent(policy, trajectory, cache)
+Agent(; policy, trajectory, env=missing) = Agent(policy, trajectory, cache)
 
 RLBase.optimise!(agent::Agent) = optimise!(TrajectoryStyle(agent.trajectory), agent)
 RLBase.optimise!(::SyncTrajectoryStyle, agent::Agent) =
@@ -49,13 +71,37 @@ end
 # in Oolong.jl with a wrapper
 function (agent::Agent)(env::AbstractEnv, args...; kwargs...)
     action = agent.policy(env, args...; kwargs...)
-    push!(agent.trajectory, (agent.cache..., action=action))
-    agent.cache = (;)
+    agent.cache.action = action
+    push!(agent.trajectory, struct_to_namedtuple(agent.cache))
+    reset!(agent.cache)
     action
 end
 
-(agent::Agent)(::PreActStage, env::AbstractEnv) =
-    agent.cache = (agent.cache..., state=state(env))
+function struct_to_namedtuple(agent_cache::AgentCache{S,R}) where {S,R}
+    attr_is_not_missing = Bool[!ismissing(getfield(agent_cache, f)) for f in cache_attrs]
+    present_attrs = cache_attrs[attr_is_not_missing]
+    return NamedTuple{present_attrs}(getfield(agent_cache, f) for f in present_attrs)
+end
 
-(agent::Agent)(::PostActStage, env::AbstractEnv) =
-    agent.cache = (agent.cache..., reward=reward(env), terminal=is_terminated(env))
+function reset!(agent_cache::AgentCache{S,R}) where {S,R}
+    agent_cache.action = missing
+    agent_cache.state = missing
+    agent_cache.reward = missing
+    agent_cache.terminal = missing
+    return nothing
+end
+
+function Base.isempty(agent_cache::AgentCache{S,R}) where {S,R}
+    ismissing(agent_cache.action) & ismissing(agent_cache.state) & ismissing(agent_cache.reward) & ismissing(agent_cache.terminal)
+end
+
+(agent::Agent)(::PreActStage, env::E) where {E <: AbstractEnv} =
+    agent.cache.state = state(env)
+    return
+
+function (agent::Agent)(::PostActStage, env::E) where {E <: AbstractEnv}
+    agent.cache.reward = reward(env)
+    agent.cache.state = state(env)
+    agent.cache.terminal = is_terminated(env)
+    return
+end