… NEWS.md
… src/ReinforcementLearningCore/src/policies/agent_srt_cache.jl
… src/ReinforcementLearningCore/src/policies/policies.jl
… src/ReinforcementLearningCore/test/core/core.jl
… src/ReinforcementLearningCore/test/policies/agent.jl
… src/ReinforcementLearningCore/test/runtests.jl