The goal is to eliminate allocations and increase the degree of type stability (not entirely type stable, but the issues are in `RLTrajectories.jl`, so not relevant for this PR.

This PR structures the `Agent` `cache` object as having `SART` fields, which makes the `cache` type stable. In the previous version, the `cache` changed type signature as different elements were included / dropped from the `NamedTuple`.

In a future version of `RLTrajectories.jl`, it would be great to be able to push a `SART` object which includes missing values, where `missingness` means that the item / key (e.g. `reward`/ `terminal` should not be added to the trajectory). That would simplify the code here significantly, but is not relevant for this PR. Timings are included below.

```julia
using BenchmarkTools
agent = Agent(
    RandomPolicy(),
    Trajectory(
        CircularArraySARTTraces(; capacity = 1_000),
        DummySampler(),
    ),
    env=RandomWalk1D() # comment this out for main branch version
)

env = RandomWalk1D()

@btime agent(env)
# Before: 937.500 ns (10 allocations: 416 bytes)
# After: 72.703 ns (0 allocations: 0 bytes)

@btime agent(PreActStage(), env)
# Before: 832.237 ns (7 allocations: 256 bytes)
# After: 42.213 ns (0 allocations: 0 bytes)

@btime agent(PostActStage(), env)
# Before: 864.583 ns (7 allocations: 320 bytes)
# After: 43.939 ns (0 allocations: 0 bytes)
```

PR Checklist

- [ ] Update NEWS.md?
- [ ] Unit tests for all structs / functions?
- [ ] Integration and correctness tests using a simple env?
- [ ] PR Review?
- [ ] Add or update documentation?
- [ ] Write docstrings for new methods?