@jeremiahpslewis 2023-03-16
@HenriDeh I don't know if the changes I made here really make sense, but if they do make sense, we get 0 allocations and a 10-20x speedup for key operations.