@HenriDeh 2023-03-16
Yeah I was think that Agent was highly type unstable with the way the cache is implemented. Give me a moment to review.