@jeremiahpslewis 2023-03-16
@HenriDeh Looking forward to your review tomorrow. I've pushed a second approach to the refactor which seems to resolve things a bit better. :)