@jeremiahpslewis 2023-03-18
@HenriDeh after three versions, think this is close to done!